Move the functions to output config file errors from scanner.c to error.c.
