+++ src/ld65/scanner.h
@@ -186,12 +186,6 @@ extern FilePos          CfgErrorPos;
 
 
 
-void CfgWarning (const FilePos* Pos, const char* Format, ...) attribute((format(printf,2,3)));
-/* Print a warning message adding file name and line number of the config file */
-
-void CfgError (const FilePos* Pos, const char* Format, ...) attribute((format(printf,2,3)));
-/* Print an error message adding file name and line number of a given file */
-
 void CfgNextTok (void);
 /* Read the next token from the input stream */
